// packages/tmdb/src/schema.d.ts
export interface paths {
// ...
  "/3/discover/movie": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Movie
     * @description Find movies using over 30 filters and sort options.
     */
    get: operations["discover-movie"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/3/discover/tv": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * TV
     * @description Find TV shows using over 30 filters and sort options.
     */
    get: operations["discover-tv"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/3/find/{external_id}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Find By ID
     * @description Find data by external ID's.
     */
    get: operations["find-by-id"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/3/genre/movie/list": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Movie List
     * @description Get the list of official genres for movies.
     */
    get: operations["genre-movie-list"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/3/genre/tv/list": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * TV List
     * @description Get the list of official genres for TV shows.
     */
    get: operations["genre-tv-list"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/movie/popular": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Popular
     * @description Get a list of movies ordered by popularity.
     */
    get: operations["movie-popular-list"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/movie/{movie_id}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Details
     * @description Get the top level details of a movie by ID.
     */
    get: operations["movie-details"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/movie/{movie_id}/recommendations": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Recommendations */
    get: operations["movie-recommendations"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/movie/{movie_id}/similar": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Similar
     * @description Get the similar movies based on genres and keywords.
     */
    get: operations["movie-similar"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/movie/{movie_id}/watch/providers": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Watch Providers
     * @description Get the list of streaming providers we have for a movie.
     */
    get: operations["movie-watch-providers"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/search/movie": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Movie
     * @description Search for movies by their original, translated and alternative titles.
     */
    get: operations["search-movie"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/3/search/multi": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Multi
     * @description Use multi search when you want to search for movies, TV shows and people in a single request.
     */
    get: operations["search-multi"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/search/tv": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * TV
     * @description Search for TV shows by their original, translated and also known as names.
     */
    get: operations["search-tv"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/3/trending/all/{time_window}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * All
     * @description Get the trending movies, TV shows and people.
     */
    get: operations["trending-all"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/3/trending/movie/{time_window}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Movies
     * @description Get the trending movies on TMDB.
     */
    get: operations["trending-movies"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/trending/tv/{time_window}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * TV
     * @description Get the trending TV shows on TMDB.
     */
    get: operations["trending-tv"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/tv/popular": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Popular
     * @description Get a list of TV shows ordered by popularity.
     */
    get: operations["tv-series-popular-list"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/tv/{series_id}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Details
     * @description Get the details of a TV show.
     */
    get: operations["tv-series-details"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/tv/{series_id}/external_ids": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * External IDs
     * @description Get a list of external IDs that have been added to a TV show.
     */
    get: operations["tv-series-external-ids"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/tv/{series_id}/recommendations": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Recommendations */
    get: operations["tv-series-recommendations"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/tv/{series_id}/similar": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Similar
     * @description Get the similar TV shows.
     */
    get: operations["tv-series-similar"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/tv/{series_id}/watch/providers": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Watch Providers
     * @description Get the list of streaming providers we have for a TV show.
     */
    get: operations["tv-series-watch-providers"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/3/tv/{series_id}/season/{season_number}": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * Details
     * @description Query the details of a TV season.
     */
    get: operations["tv-season-details"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: never;
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
// ...
  "account-get-favorites": {
    parameters: {
      query?: {
        language?: string;
        page?: number;
        session_id?: string;
        sort_by?: "created_at.asc" | "created_at.desc";
      };
      header?: never;
      path: {
        account_id: number;
      };
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 200 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": {
            /**
             * @default 0
             * @example 1
             */
            page: number;
            results?: {
              /**
               * @default true
               * @example false
               */
              adult: boolean;
              /** @example /se5Hxz7PArQZOG3Nx2bpfOhLhtV.jpg */
              backdrop_path?: string;
              genre_ids?: number[];
              /**
               * @default 0
               * @example 9806
               */
              id: number;
              /** @example en */
              original_language?: string;
              /** @example The Incredibles */
              original_title?: string;
              /** @example Bob Parr has given up his superhero days to log in time as an insurance adjuster and raise his three children with his formerly heroic wife in suburbia. But when he receives a mysterious assignment, it's time to get back into costume. */
              overview?: string;
              /**
               * @default 0
               * @example 71.477
               */
              popularity: number;
              /** @example /2LqaLgk4Z226KkgPJuiOQ58wvrm.jpg */
              poster_path?: string;
              /** @example 2004-10-27 */
              release_date?: string;
              /** @example The Incredibles */
              title?: string;
              /**
               * @default true
               * @example false
               */
              video: boolean;
              /**
               * @default 0
               * @example 7.702
               */
              vote_average: number;
              /**
               * @default 0
               * @example 16162
               */
              vote_count: number;
            }[];
            /**
             * @default 0
             * @example 4
             */
            total_pages: number;
            /**
             * @default 0
             * @example 80
             */
            total_results: number;
          };
        };
      };
    };
  };
// ...
  "certification-movie-list": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 200 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": {
            certifications?: {
// ...
              US?: {
                /** @example R */
                certification?: string;
                /** @example Under 17 requires accompanying parent or adult guardian 21 or older. The parent/guardian is required to stay with the child under 17 through the entire movie, even if the parent gives the child/teenager permission to see the film alone. These films may contain strong profanity, graphic sexuality, nudity, strong violence, horror, gore, and strong drug use. A movie rated R for profanity often has more severe or frequent language than the PG-13 rating would permit. An R-rated movie may have more blood, gore, drug use, nudity, or graphic sexuality than a PG-13 movie would admit. */
                meaning?: string;
                /**
                 * @default 0
                 * @example 4
                 */
                order: number;
              }[];
// ...
            };
          };
        };
      };
    };
  };
// ...
  "genre-movie-list": {
    parameters: {
      query?: {
        language?: string;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 200 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": {
            genres?: {
              /**
               * @default 0
               * @example 28
               */
              id: number;
              /** @example Action */
              name?: string;
            }[];
          };
        };
      };
    };
  };
// ...
}